# Swap loops and improve segmented analysis